separated getters from impl, added global getter errors
